"#11": Errores mostrados en los inputs
Se han modificado los inputs para mostrar los errores, así como iluminarse en azul al presionarlos
