+++ Prueba.js
@@ -4,7 +4,7 @@ import {Block} from 'galio-framework';
 import FloatingLabelInput from './components/FloatingLabelInput';
 import { Icon } from './components';
 import { argonTheme } from './constants';
-import FormErrorMessage from './components/FormErrorMessage';
+
 
 
 export default function Prueba({navigation}) {
@@ -21,28 +21,7 @@ export default function Prueba({navigation}) {
       <SafeAreaView >
       <ScrollView  showsVerticalScrollIndicator={false} keyboardShouldPersistTaps={true}>
       
-      <Block> 
-      
-    
-                     <FloatingLabelInput
-                      label="Emaila"
-                      onChangeText={text => setForm({...form,["email"]:text.replace(/\s/g, "")})}
-                      iconContent={<Icon
-                        
-                        size={20}
-                        color={argonTheme.COLORS.ICON}
-                        name="ic_mail_24px"
-                        family="ArgonExtra"
-                        style={{marginRight: 8}}
-                      />}
-                      
-                     
-                    />
-                    
-                  
-                    
      
-      </Block>
       </ScrollView>
       </SafeAreaView>
     )
