+++ components/Delay.js
@@ -1,0 +1,7 @@
+
+const delay = ms => new Promise(
+    resolve => setTimeout(resolve, ms)
+  );
+
+export {delay}
+
