+++ constants/Theme.js
@@ -27,6 +27,7 @@ export default {
     HEADER: '#525F7F',
     BORDER: '#CAD1D7',
     WHITE: '#FFFFFF',
-    BLACK: '#000000'
+    BLACK: '#000000',
+    MESSAGE_ERROR: '#E82825',
   }
 };
